Refactor buffs
