--- rpg.Game/Buffs/BuffType.cs
@@ -1,7 +1,0 @@
-﻿namespace rpg.Game.Buffs
-{
-    public enum BuffType
-    {
-        XPBoost, HealthBoost, ManaBoost, StaminaBoost, Drawing
-    }
-}
