+++ rpg.Game/Effects/EffectType.cs
@@ -1,0 +1,11 @@
+﻿namespace rpg.Game.Effects
+{
+    public enum BuffType
+    {
+        XPBoost, 
+        HealthBoost, 
+        ManaBoost, 
+        StaminaBoost, 
+        Drawing
+    }
+}
