+++ rpg.Game/Effects/XPBoost.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace rpg.Game.Buffs
+namespace rpg.Game.Effects
 {
-    public class XPBoost : Buff
+    public class XPBoost : Effect
     {
         public XPBoost(int duration)
         {
@@ -16,7 +16,6 @@ namespace rpg.Game.Buffs
         public override void OnTick()
         {
             base.OnTick();
-
         }
     }
 }
