+++ rpg.Game/Entities/Entity.cs
@@ -35,12 +35,13 @@ namespace rpg.Game.Entities
         public event EventHandler<EventArgs> DrawOrderChanged;
         public event EventHandler<EventArgs> VisibleChanged;
 
+        public abstract void SetUp();
+
         public void Draw(GameTime gameTime)
         { 
             model.Draw(new Matrix(), new Matrix(), new Matrix());
         }
 
         public abstract void Update(GameTime gameTime);
-
     }
 }
